jmap_quota.c: create separate source file for Quota/ methods
