+++ imap/http_jmap.c
@@ -186,6 +186,7 @@ static void jmap_init(struct buf *serverinfo)
 
     jmap_core_init(&my_jmap_settings);
     jmap_blob_init(&my_jmap_settings);
+    jmap_quota_init(&my_jmap_settings);
     jmap_mail_init(&my_jmap_settings);
     jmap_mdn_init(&my_jmap_settings);
     jmap_contact_init(&my_jmap_settings);
