+++ imap/jmap_api.c
@@ -556,6 +556,7 @@ static json_t *lookup_capabilities(const char *accountid,
         /* Primary account has all capabilities */
         jmap_core_capabilities(capas);
         jmap_blob_capabilities(capas);
+        jmap_quota_capabilities(capas);
         jmap_mail_capabilities(capas, mayCreateTopLevel);
         jmap_emailsubmission_capabilities(capas);
         jmap_mdn_capabilities(capas);
@@ -580,6 +581,7 @@ static json_t *lookup_capabilities(const char *accountid,
         if (rock.is_visible) {
             jmap_core_capabilities(capas);
             jmap_blob_capabilities(capas);
+            jmap_quota_capabilities(capas);
             if (rock.has_mail) {
                 // we don't offer emailsubmission or vacation
                 // for shared accounts right now
